add check removable for deleteEquipment, fixed the bug of forevering loading when no equipment is found

+++ LabMan-Backend/src/controllers/equipment/deleteEquipment.js
@@ -1,20 +1,44 @@
 import pool from "../../utils/MySQL/db.js";
 
-function deleteEquipment(req, res) {
+async function deleteEquipment(req, res) {
 	const type_id = req.params.type_id;
-	pool.query(
-		"DELETE FROM equipment_type WHERE type_id = ?",
-		[type_id],
-		(err, results) => {
-			if (err) {
-				return res.status(500).json({ message: "Cannot delete or update this equipment type as it is currently in use by other records.", errorDetails: err.message });
-			}
-			if (results.affectedRows === 0) {
-				return res.status(404).json({ error: "Equipment not found" });
-			}
-			return res.status(200).json({ message: "Equipment deleted successfully" });
+
+	checkEquipmentRemovable(type_id, (err, removable) => {
+		if (err && err.message === "Equipment type not found") {
+			res.status(404).json({ error: "Equipment type not found" });
+		}else if (err) {
+			res.status(500).json({ error: err.message });
+		}else if (removable === 0) {
+			res.status(400).json({error: "Equipment is not removable" });
+		}else{
+			const query = "DELETE FROM equipment_type WHERE type_id = ?";
+			const params = [type_id];
+
+			pool.query(query, params, (err) => {
+				if (err) {
+					res.status(500).json({ error: "Error deleting equipment" });
+				}else{
+					res.status(200).json({ message: "Equipment deleted" });
+				}
+			});
 		}
-	);
+	});
 }
+
+const checkEquipmentRemovable = (type_id, callback) => {
+	const query = "SELECT removable FROM equipment_type WHERE type_id = ?";
+	const params = [type_id];
+
+	pool.query(query, params, (err, results) => {
+		if (err) {
+			callback(new Error("Error checking equipment removable"));
+		}else if (results.length === 0) {
+			callback(new Error("Equipment type not found"));
+		}else{
+			const removable = results[0].removable;
+			callback(null, removable);
+		}
+	});
+};
 
 export { deleteEquipment };
